+++ backend/docker/scripts/restore-backup.sh
@@ -9,16 +9,17 @@ if [ ! -f substrate-data.tar ]; then
 fi
 
 # Stop all cointainers
+docker stop backend_api_1
 docker stop backend_crawler_1
 docker stop backend_graphql-engine_1
 docker stop backend_postgres_1
 docker stop backend_substrate-node_1
 sleep 5
 
-if [ $(docker inspect -f '{{.State.Running}}' backend_substrate-node_1) = "true" ]; then 
+if [ $(docker inspect -f '{{.State.Running}}' backend_substrate-node_1) = "true" ]; then
     echo "Substrate node still running"
     exit 1
-else 
+else
     echo "Substrate node stopped"
     echo "Starting the restoration"
 fi
@@ -35,5 +36,6 @@ docker start backend_substrate-node_1
 docker start backend_postgres_1
 docker start backend_graphql-engine_1
 docker start backend_crawler_1
+docker start backend_api_1