+++ frontend/CHANGELOG.md
@@ -5,3 +5,5 @@
 - Updated header component
 - Removed github corner element
 - Removed additional pages
+- Updated footer component
+- Updated title and favicon